Issue #1555: Extract classes to the upper level
Fixes `PackageVisibleInnerClass` inspection violations.

Description:
>Reports package-local inner classes.
