+++ src/main/java/com/puppycrawl/tools/checkstyle/gui/Listener.java
@@ -1,0 +1,47 @@
+////////////////////////////////////////////////////////////////////////////////
+// checkstyle: Checks Java source code for adherence to a set of rules.
+// Copyright (C) 2001-2015 the original author or authors.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+////////////////////////////////////////////////////////////////////////////////
+
+package com.puppycrawl.tools.checkstyle.gui;
+
+import java.io.File;
+
+/**
+ * Implement this inner interface to listen for when files are dropped. For example
+ * your class declaration may begin like this:
+ * {@code <pre>
+ *      public class MyClass implements FileDrop.Listener
+ *      ...
+ *      public void filesDropped( File[] files )
+ *      {
+ *          ...
+ *      }   // end filesDropped
+ *      ...
+ * </pre>}
+ *
+ * @author Lars Kühne
+ */
+interface Listener {
+    /**
+     * This method is called when files have been successfully dropped.
+     *
+     * @param files An array of <tt>File</tt>s that were dropped.
+     * @since 1.0
+     */
+    void filesDropped(File... files);
+}
